+++ qt/python/mantidqt/mantidqt/widgets/codeeditor/interpreter.py
@@ -10,25 +10,35 @@ import sys
 import traceback
 
 from qtpy.QtCore import QObject, Qt, Signal
-from qtpy.QtGui import QFont, QFontMetrics
+from qtpy.QtGui import QFont, QFontMetrics, QColor
 from qtpy.QtWidgets import QFileDialog, QMessageBox, QStatusBar, QVBoxLayout, QWidget
 
 from mantidqt.io import open_a_file_dialog
 from mantid.kernel import config
+import mantid.kernel.environment as mtd_env
 from mantidqt.widgets.codeeditor.codecommenter import CodeCommenter
 from mantidqt.widgets.codeeditor.completion import CodeCompleter
 from mantidqt.widgets.codeeditor.editor import CodeEditor
 from mantidqt.widgets.codeeditor.errorformatter import ErrorFormatter
 from mantidqt.widgets.codeeditor.execution import PythonCodeExecution
 from mantidqt.widgets.embedded_find_replace_dialog.presenter import EmbeddedFindReplaceDialog
-from workbench.config.fonts import CURRENTLINE_BKGD_COLOR
 
 IDLE_STATUS_MSG = "Status: Idle."
 LAST_JOB_MSG_TEMPLATE = "Last job completed {} at {} in {:.3f}s"
 RUNNING_STATUS_MSG = "Status: Running"
 ABORTED_STATUS_MSG = "Status: Aborted"
 
+
+# Returns the background color of the current line in the code editor
+def get_currentline_background_color():
+    if mtd_env.is_mac():
+        if mtd_env.is_dark_mode():
+            return QColor(0, 52, 110)
+    return QColor(247, 236, 248)
+
+
 # Editor
+CURRENTLINE_BKGD_COLOR = get_currentline_background_color()
 TAB_WIDTH = 4
 TAB_CHAR = "\t"
 SPACE_CHAR = " "
